7/14 17:05
完成了定期檢查團購狀態、判斷人數以及成立訂單的兩支排程器
接下來是後台查詢訂單的頁面

+++ GroupBuy/src/com/gro_order/controller/Gro_orderServlet.java
@@ -72,7 +72,7 @@ public class Gro_orderServlet extends HttpServlet {
 				Timestamp ord_date = gro_orderVO.getOrd_date();
 				String mem_id = gro_orderVO.getMem_id();
 				String gro_id = gro_orderVO.getGro_id();
-				Integer ord_price = gro_orderVO.getOrd_price();
+				Double ord_price = gro_orderVO.getOrd_price();
 				String receive_name = gro_orderVO.getReceive_name();
 				String address = gro_orderVO.getAddress();
 				String phone = gro_orderVO.getPhone();
